Cleanups for glTF extensions support
